vid2mp3py/src/authorization/requirements.txt
pyjwt
flask
flask_mysqldb